requirements.txt
streamlit
google-generativeai
pypdf
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy